[test_loop] Refactor all test loop tests into a separate module
