+++ integration-tests/src/lib.rs
@@ -1,10 +1,8 @@
-pub mod genesis_helpers;
-pub mod nearcore_utils;
 pub mod node;
 pub mod runtime_utils;
-pub mod test_helpers;
 pub mod user;
 
-pub mod test_loop;
+#[cfg(test)]
+mod test_loop;
 #[cfg(test)]
 mod tests;
