+++ integration-tests/src/test_loop/mod.rs
@@ -1,2 +1,3 @@
-pub mod builder;
-pub mod env;
+mod builder;
+mod env;
+mod tests;
