+++ integration-tests/src/test_loop/tests/mod.rs
@@ -1,0 +1,4 @@
+pub mod in_memory_tries;
+pub mod multinode_stateless_validators;
+pub mod multinode_test_loop_example;
+pub mod simple_test_loop_example;
