+++ integration-tests/src/tests/mod.rs
@@ -8,3 +8,5 @@ mod test_errors;
 mod test_overflows;
 mod test_simple;
 mod test_tps_regression;
+mod test_helpers;
+mod genesis_helpers;
