+++ integration-tests/src/tests/nearcore/rpc_nodes.rs
@@ -1,4 +1,4 @@
-use crate::genesis_helpers::genesis_block;
+use crate::tests::genesis_helpers::genesis_block;
 use crate::tests::nearcore::node_cluster::NodeCluster;
 use actix::clock::sleep;
 use actix::{Actor, System};
