+++ integration-tests/src/tests/standard_cases/rpc.rs
@@ -2,8 +2,8 @@
 //! The communication is performed through `RPCUser` that uses the standard RPC API to communicate.
 
 use crate::node::{create_nodes_from_seeds, Node, NodeConfig, ThreadNode};
-use crate::test_helpers::heavy_test;
 use crate::tests::standard_cases::*;
+use crate::tests::test_helpers::heavy_test;
 use near_o11y::testonly::init_test_module_logger;
 use std::thread;
 use std::time::Duration;
